deprecate MoveGroup class and introduce MoveGroupInterface
The name of the class is confusing to many users and many people
do not even know that the class relies entirely on ROS interfaces
and does not implement much logic on its own.
