+++ moveit_ros/planning_interface/move_group_interface/src/move_group_interface.cpp
@@ -42,7 +42,7 @@
 #include <moveit/kinematic_constraints/utils.h>
 #include <moveit/move_group/capability_names.h>
 #include <moveit/move_group_pick_place_capability/capability_names.h>
-#include <moveit/move_group_interface/move_group.h>
+#include <moveit/move_group_interface/move_group_interface.h>
 #include <moveit/planning_scene_monitor/current_state_monitor.h>
 #include <moveit/planning_scene_monitor/planning_scene_monitor.h>
 #include <moveit/trajectory_execution_manager/trajectory_execution_manager.h>
@@ -72,7 +72,7 @@ namespace moveit
 namespace planning_interface
 {
 
-const std::string MoveGroup::ROBOT_DESCRIPTION = "robot_description";    // name of the robot description (a param name, so it can be changed externally)
+const std::string MoveGroupInterface::ROBOT_DESCRIPTION = "robot_description";    // name of the robot description (a param name, so it can be changed externally)
 
 namespace
 {
@@ -84,11 +84,11 @@ enum ActiveTargetType
 
 }
 
-class MoveGroup::MoveGroupImpl
+class MoveGroupInterface::MoveGroupInterfaceImpl
 {
 public:
 
-  MoveGroupImpl(const Options &opt, const boost::shared_ptr<tf::Transformer> &tf, const ros::WallDuration &wait_for_servers)
+  MoveGroupInterfaceImpl(const Options &opt, const boost::shared_ptr<tf::Transformer> &tf, const ros::WallDuration &wait_for_servers)
     : opt_(opt),
       node_handle_(opt.node_handle_),
       tf_(tf)
@@ -163,13 +163,13 @@ public:
 
     cartesian_path_service_ = node_handle_.serviceClient<moveit_msgs::GetCartesianPath>(move_group::CARTESIAN_PATH_SERVICE_NAME);
 
-    ROS_INFO_STREAM("Ready to take MoveGroup commands for group " << opt.group_name_ << ".");
+    ROS_INFO_STREAM("Ready to take commands for planning group " << opt.group_name_ << ".");
   }
 
   template<typename T>
   void waitForAction(const T &action, const std::string &name, const ros::WallTime &timeout, double allotted_time)
   {
-    ROS_DEBUG("Waiting for MoveGroup action server (%s)...", name.c_str());
+    ROS_DEBUG("Waiting for move_group action server (%s)...", name.c_str());
 
     // wait for the server (and spin as needed)
     if (timeout == ros::WallTime())
@@ -230,7 +230,7 @@ public:
     }
   }
 
-  ~MoveGroupImpl()
+  ~MoveGroupInterfaceImpl()
   {
     if (constraints_init_thread_)
       constraints_init_thread_->join();
@@ -1093,7 +1093,7 @@ public:
     initializing_constraints_ = true;
     if (constraints_init_thread_)
       constraints_init_thread_->join();
-    constraints_init_thread_.reset(new boost::thread(boost::bind(&MoveGroupImpl::initializeConstraintsStorageThread, this, host, port)));
+    constraints_init_thread_.reset(new boost::thread(boost::bind(&MoveGroupInterfaceImpl::initializeConstraintsStorageThread, this, host, port)));
   }
 
   void setWorkspace(double minx, double miny, double minz, double maxx, double maxy, double maxz)
@@ -1184,40 +1184,40 @@ private:
 }
 }
 
-moveit::planning_interface::MoveGroup::MoveGroup(const std::string &group_name, const boost::shared_ptr<tf::Transformer> &tf, const ros::WallDuration &wait_for_servers)
+moveit::planning_interface::MoveGroupInterface::MoveGroupInterface(const std::string &group_name, const boost::shared_ptr<tf::Transformer> &tf, const ros::WallDuration &wait_for_servers)
 {
   if (!ros::ok())
     throw std::runtime_error("ROS does not seem to be running");
-  impl_ = new MoveGroupImpl(Options(group_name), tf ? tf : getSharedTF(), wait_for_servers);
+  impl_ = new MoveGroupInterfaceImpl(Options(group_name), tf ? tf : getSharedTF(), wait_for_servers);
 }
 
-moveit::planning_interface::MoveGroup::MoveGroup(const std::string &group, const boost::shared_ptr<tf::Transformer> &tf, const ros::Duration &wait_for_servers)
-  : MoveGroup(group, tf, ros::WallDuration(wait_for_servers.toSec()))
+moveit::planning_interface::MoveGroupInterface::MoveGroupInterface(const std::string &group, const boost::shared_ptr<tf::Transformer> &tf, const ros::Duration &wait_for_servers)
+  : MoveGroupInterface(group, tf, ros::WallDuration(wait_for_servers.toSec()))
 {
 }
 
-moveit::planning_interface::MoveGroup::MoveGroup(const Options &opt, const boost::shared_ptr<tf::Transformer> &tf, const ros::WallDuration &wait_for_servers)
+moveit::planning_interface::MoveGroupInterface::MoveGroupInterface(const Options &opt, const boost::shared_ptr<tf::Transformer> &tf, const ros::WallDuration &wait_for_servers)
 {
-  impl_ = new MoveGroupImpl(opt, tf ? tf : getSharedTF(), wait_for_servers);
+  impl_ = new MoveGroupInterfaceImpl(opt, tf ? tf : getSharedTF(), wait_for_servers);
 }
 
-moveit::planning_interface::MoveGroup::MoveGroup(const moveit::planning_interface::MoveGroup::Options &opt,
+moveit::planning_interface::MoveGroupInterface::MoveGroupInterface(const moveit::planning_interface::MoveGroupInterface::Options &opt,
                                                  const boost::shared_ptr<tf::Transformer> &tf, const ros::Duration &wait_for_servers)
-  : MoveGroup(opt, tf, ros::WallDuration(wait_for_servers.toSec()))
+  : MoveGroupInterface(opt, tf, ros::WallDuration(wait_for_servers.toSec()))
 {
 }
 
-moveit::planning_interface::MoveGroup::~MoveGroup()
+moveit::planning_interface::MoveGroupInterface::~MoveGroupInterface()
 {
   delete impl_;
 }
 
-const std::string& moveit::planning_interface::MoveGroup::getName() const
+const std::string& moveit::planning_interface::MoveGroupInterface::getName() const
 {
   return impl_->getOptions().group_name_;
 }
 
-const std::vector<std::string> moveit::planning_interface::MoveGroup::getNamedTargets()
+const std::vector<std::string> moveit::planning_interface::MoveGroupInterface::getNamedTargets()
 {
   const robot_model::RobotModelConstPtr& robot = getRobotModel();
   const std::string& group = getName();
@@ -1232,121 +1232,121 @@ const std::vector<std::string> moveit::planning_interface::MoveGroup::getNamedTa
   return empty;
 }
 
-robot_model::RobotModelConstPtr moveit::planning_interface::MoveGroup::getRobotModel() const
+robot_model::RobotModelConstPtr moveit::planning_interface::MoveGroupInterface::getRobotModel() const
 {
   return impl_->getRobotModel();
 }
 
-const ros::NodeHandle& moveit::planning_interface::MoveGroup::getNodeHandle() const
+const ros::NodeHandle& moveit::planning_interface::MoveGroupInterface::getNodeHandle() const
 {
   return impl_->getOptions().node_handle_;
 }
 
-bool moveit::planning_interface::MoveGroup::getInterfaceDescription(moveit_msgs::PlannerInterfaceDescription &desc)
+bool moveit::planning_interface::MoveGroupInterface::getInterfaceDescription(moveit_msgs::PlannerInterfaceDescription &desc)
 {
   return impl_->getInterfaceDescription(desc);
 }
 
 std::map<std::string, std::string>
-moveit::planning_interface::MoveGroup::getPlannerParams(const std::string &planner_id,
+moveit::planning_interface::MoveGroupInterface::getPlannerParams(const std::string &planner_id,
                                                         const std::string &group) {
   return impl_->getPlannerParams(planner_id, group);
 }
 
 
-void moveit::planning_interface::MoveGroup::setPlannerParams(const std::string &planner_id,
+void moveit::planning_interface::MoveGroupInterface::setPlannerParams(const std::string &planner_id,
                                                              const std::string &group,
                                                              const std::map<std::string, std::string> &params,
                                                              bool replace) {
   impl_->setPlannerParams(planner_id, group, params, replace);
 }
 
-std::string moveit::planning_interface::MoveGroup::getDefaultPlannerId(const std::string &group) const
+std::string moveit::planning_interface::MoveGroupInterface::getDefaultPlannerId(const std::string &group) const
 {
   return impl_->getDefaultPlannerId(group);
 }
 
-void moveit::planning_interface::MoveGroup::setPlannerId(const std::string &planner_id)
+void moveit::planning_interface::MoveGroupInterface::setPlannerId(const std::string &planner_id)
 {
   impl_->setPlannerId(planner_id);
 }
 
-void moveit::planning_interface::MoveGroup::setNumPlanningAttempts(unsigned int num_planning_attempts)
+void moveit::planning_interface::MoveGroupInterface::setNumPlanningAttempts(unsigned int num_planning_attempts)
 {
   impl_->setNumPlanningAttempts(num_planning_attempts);
 }
 
-void moveit::planning_interface::MoveGroup::setMaxVelocityScalingFactor(double max_velocity_scaling_factor)
+void moveit::planning_interface::MoveGroupInterface::setMaxVelocityScalingFactor(double max_velocity_scaling_factor)
 {
   impl_->setMaxVelocityScalingFactor(max_velocity_scaling_factor);
 }
 
-void moveit::planning_interface::MoveGroup::setMaxAccelerationScalingFactor(double max_acceleration_scaling_factor)
+void moveit::planning_interface::MoveGroupInterface::setMaxAccelerationScalingFactor(double max_acceleration_scaling_factor)
 {
   impl_->setMaxAccelerationScalingFactor(max_acceleration_scaling_factor);
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::asyncMove()
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::asyncMove()
 {
   return impl_->move(false);
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::move()
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::move()
 {
   return impl_->move(true);
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::asyncExecute(const Plan &plan)
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::asyncExecute(const Plan &plan)
 {
   return impl_->execute(plan, false);
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::execute(const Plan &plan)
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::execute(const Plan &plan)
 {
   return impl_->execute(plan, true);
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::plan(Plan &plan)
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::plan(Plan &plan)
 {
   return impl_->plan(plan);
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::pick(const std::string &object)
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::pick(const std::string &object)
 {
   return impl_->pick(object, std::vector<moveit_msgs::Grasp>());
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::pick(const std::string &object, const moveit_msgs::Grasp &grasp)
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::pick(const std::string &object, const moveit_msgs::Grasp &grasp)
 {
   return impl_->pick(object, std::vector<moveit_msgs::Grasp>(1, grasp));
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::pick(const std::string &object, const std::vector<moveit_msgs::Grasp> &grasps)
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::pick(const std::string &object, const std::vector<moveit_msgs::Grasp> &grasps)
 {
   return impl_->pick(object, grasps);
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::place(const std::string &object)
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::place(const std::string &object)
 {
   return impl_->place(object, std::vector<moveit_msgs::PlaceLocation>());
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::place(const std::string &object, const std::vector<moveit_msgs::PlaceLocation> &locations)
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::place(const std::string &object, const std::vector<moveit_msgs::PlaceLocation> &locations)
 {
   return impl_->place(object, locations);
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::place(const std::string &object, const std::vector<geometry_msgs::PoseStamped> &poses)
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::place(const std::string &object, const std::vector<geometry_msgs::PoseStamped> &poses)
 {
   return impl_->place(object, poses);
 }
 
-moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroup::place(const std::string &object, const geometry_msgs::PoseStamped &pose)
+moveit::planning_interface::MoveItErrorCode moveit::planning_interface::MoveGroupInterface::place(const std::string &object, const geometry_msgs::PoseStamped &pose)
 {
   return impl_->place(object, std::vector<geometry_msgs::PoseStamped>(1, pose));
 }
 
-double moveit::planning_interface::MoveGroup::computeCartesianPath(const std::vector<geometry_msgs::Pose> &waypoints, double eef_step, double jump_threshold,
+double moveit::planning_interface::MoveGroupInterface::computeCartesianPath(const std::vector<geometry_msgs::Pose> &waypoints, double eef_step, double jump_threshold,
                                                                    moveit_msgs::RobotTrajectory &trajectory, bool avoid_collisions,
 								   moveit_msgs::MoveItErrorCodes *error_code)
 {
@@ -1354,7 +1354,7 @@ double moveit::planning_interface::MoveGroup::computeCartesianPath(const std::ve
   return computeCartesianPath(waypoints, eef_step, jump_threshold, trajectory, path_constraints_tmp, avoid_collisions, error_code);
 }
 
-double moveit::planning_interface::MoveGroup::computeCartesianPath(const std::vector<geometry_msgs::Pose> &waypoints, double eef_step, double jump_threshold,
+double moveit::planning_interface::MoveGroupInterface::computeCartesianPath(const std::vector<geometry_msgs::Pose> &waypoints, double eef_step, double jump_threshold,
                                                                    moveit_msgs::RobotTrajectory &trajectory,
                                                                    const moveit_msgs::Constraints &path_constraints, bool avoid_collisions,
                                                                    moveit_msgs::MoveItErrorCodes *error_code)
@@ -1370,12 +1370,12 @@ double moveit::planning_interface::MoveGroup::computeCartesianPath(const std::ve
   }
 }
 
-void moveit::planning_interface::MoveGroup::stop()
+void moveit::planning_interface::MoveGroupInterface::stop()
 {
   impl_->stop();
 }
 
-void moveit::planning_interface::MoveGroup::setStartState(const moveit_msgs::RobotState &start_state)
+void moveit::planning_interface::MoveGroupInterface::setStartState(const moveit_msgs::RobotState &start_state)
 {
   robot_state::RobotStatePtr rs;
   impl_->getCurrentState(rs);
@@ -1383,28 +1383,28 @@ void moveit::planning_interface::MoveGroup::setStartState(const moveit_msgs::Rob
   setStartState(*rs);
 }
 
-void moveit::planning_interface::MoveGroup::setStartState(const robot_state::RobotState &start_state)
+void moveit::planning_interface::MoveGroupInterface::setStartState(const robot_state::RobotState &start_state)
 {
   impl_->setStartState(start_state);
 }
 
-void moveit::planning_interface::MoveGroup::setStartStateToCurrentState()
+void moveit::planning_interface::MoveGroupInterface::setStartStateToCurrentState()
 {
   impl_->setStartStateToCurrentState();
 }
 
-void moveit::planning_interface::MoveGroup::setRandomTarget()
+void moveit::planning_interface::MoveGroupInterface::setRandomTarget()
 {
   impl_->getJointStateTarget().setToRandomPositions();
   impl_->setTargetType(JOINT);
 }
 
-const std::vector<std::string>& moveit::planning_interface::MoveGroup::getJointNames()
+const std::vector<std::string>& moveit::planning_interface::MoveGroupInterface::getJointNames()
 {
   return impl_->getJointModelGroup()->getVariableNames();
 }
 
-std::map<std::string, double> moveit::planning_interface::MoveGroup::getNamedTargetValues(const std::string& name)
+std::map<std::string, double> moveit::planning_interface::MoveGroupInterface::getNamedTargetValues(const std::string& name)
 {
   std::map<std::string, std::vector<double> >::const_iterator it = remembered_joint_values_.find(name);
   std::map<std::string,double> positions;
@@ -1424,7 +1424,7 @@ std::map<std::string, double> moveit::planning_interface::MoveGroup::getNamedTar
   return positions;
 }
 
-bool moveit::planning_interface::MoveGroup::setNamedTarget(const std::string &name)
+bool moveit::planning_interface::MoveGroupInterface::setNamedTarget(const std::string &name)
 {
   std::map<std::string, std::vector<double> >::const_iterator it = remembered_joint_values_.find(name);
   if (it != remembered_joint_values_.end())
@@ -1442,7 +1442,7 @@ bool moveit::planning_interface::MoveGroup::setNamedTarget(const std::string &na
   }
 }
 
-bool moveit::planning_interface::MoveGroup::setJointValueTarget(const std::vector<double> &joint_values)
+bool moveit::planning_interface::MoveGroupInterface::setJointValueTarget(const std::vector<double> &joint_values)
 {
   if (joint_values.size() != impl_->getJointModelGroup()->getVariableCount())
     return false;
@@ -1451,27 +1451,27 @@ bool moveit::planning_interface::MoveGroup::setJointValueTarget(const std::vecto
   return impl_->getJointStateTarget().satisfiesBounds(impl_->getJointModelGroup(), impl_->getGoalJointTolerance());
 }
 
-bool moveit::planning_interface::MoveGroup::setJointValueTarget(const std::map<std::string, double> &joint_values)
+bool moveit::planning_interface::MoveGroupInterface::setJointValueTarget(const std::map<std::string, double> &joint_values)
 {
   impl_->setTargetType(JOINT);
   impl_->getJointStateTarget().setVariablePositions(joint_values);
   return impl_->getJointStateTarget().satisfiesBounds(impl_->getGoalJointTolerance());
 }
 
-bool moveit::planning_interface::MoveGroup::setJointValueTarget(const robot_state::RobotState &rstate)
+bool moveit::planning_interface::MoveGroupInterface::setJointValueTarget(const robot_state::RobotState &rstate)
 {
   impl_->setTargetType(JOINT);
   impl_->getJointStateTarget() = rstate;
   return impl_->getJointStateTarget().satisfiesBounds(impl_->getGoalJointTolerance());
 }
 
-bool moveit::planning_interface::MoveGroup::setJointValueTarget(const std::string &joint_name, double value)
+bool moveit::planning_interface::MoveGroupInterface::setJointValueTarget(const std::string &joint_name, double value)
 {
   std::vector<double> values(1, value);
   return setJointValueTarget(joint_name, values);
 }
 
-bool moveit::planning_interface::MoveGroup::setJointValueTarget(const std::string &joint_name, const std::vector<double> &values)
+bool moveit::planning_interface::MoveGroupInterface::setJointValueTarget(const std::string &joint_name, const std::vector<double> &values)
 {
   impl_->setTargetType(JOINT);
   const robot_model::JointModel *jm = impl_->getJointStateTarget().getJointModel(joint_name);
@@ -1483,63 +1483,63 @@ bool moveit::planning_interface::MoveGroup::setJointValueTarget(const std::strin
   return false;
 }
 
-bool moveit::planning_interface::MoveGroup::setJointValueTarget(const sensor_msgs::JointState &state)
+bool moveit::planning_interface::MoveGroupInterface::setJointValueTarget(const sensor_msgs::JointState &state)
 {
   impl_->setTargetType(JOINT);
   impl_->getJointStateTarget().setVariableValues(state);
   return impl_->getJointStateTarget().satisfiesBounds(impl_->getGoalJointTolerance());
 }
 
-bool moveit::planning_interface::MoveGroup::setJointValueTarget(const geometry_msgs::Pose &eef_pose, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setJointValueTarget(const geometry_msgs::Pose &eef_pose, const std::string &end_effector_link)
 {
   return impl_->setJointValueTarget(eef_pose, end_effector_link, "", false);
 }
 
-bool moveit::planning_interface::MoveGroup::setJointValueTarget(const geometry_msgs::PoseStamped &eef_pose, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setJointValueTarget(const geometry_msgs::PoseStamped &eef_pose, const std::string &end_effector_link)
 {
   return impl_->setJointValueTarget(eef_pose.pose, end_effector_link, eef_pose.header.frame_id, false);
 }
 
-bool moveit::planning_interface::MoveGroup::setJointValueTarget(const Eigen::Affine3d &eef_pose, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setJointValueTarget(const Eigen::Affine3d &eef_pose, const std::string &end_effector_link)
 {
   geometry_msgs::Pose msg;
   tf::poseEigenToMsg(eef_pose, msg);
   return setJointValueTarget(msg, end_effector_link);
 }
 
-bool moveit::planning_interface::MoveGroup::setApproximateJointValueTarget(const geometry_msgs::Pose &eef_pose, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setApproximateJointValueTarget(const geometry_msgs::Pose &eef_pose, const std::string &end_effector_link)
 {
   return impl_->setJointValueTarget(eef_pose, end_effector_link, "", true);
 }
 
-bool moveit::planning_interface::MoveGroup::setApproximateJointValueTarget(const geometry_msgs::PoseStamped &eef_pose, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setApproximateJointValueTarget(const geometry_msgs::PoseStamped &eef_pose, const std::string &end_effector_link)
 {
   return impl_->setJointValueTarget(eef_pose.pose, end_effector_link, eef_pose.header.frame_id, true);
 }
 
-bool moveit::planning_interface::MoveGroup::setApproximateJointValueTarget(const Eigen::Affine3d &eef_pose, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setApproximateJointValueTarget(const Eigen::Affine3d &eef_pose, const std::string &end_effector_link)
 {
   geometry_msgs::Pose msg;
   tf::poseEigenToMsg(eef_pose, msg);
   return setApproximateJointValueTarget(msg, end_effector_link);
 }
 
-const robot_state::RobotState& moveit::planning_interface::MoveGroup::getJointValueTarget() const
+const robot_state::RobotState& moveit::planning_interface::MoveGroupInterface::getJointValueTarget() const
 {
   return impl_->getJointStateTarget();
 }
 
-const std::string& moveit::planning_interface::MoveGroup::getEndEffectorLink() const
+const std::string& moveit::planning_interface::MoveGroupInterface::getEndEffectorLink() const
 {
   return impl_->getEndEffectorLink();
 }
 
-const std::string& moveit::planning_interface::MoveGroup::getEndEffector() const
+const std::string& moveit::planning_interface::MoveGroupInterface::getEndEffector() const
 {
   return impl_->getEndEffector();
 }
 
-bool moveit::planning_interface::MoveGroup::setEndEffectorLink(const std::string &link_name)
+bool moveit::planning_interface::MoveGroupInterface::setEndEffectorLink(const std::string &link_name)
 {
   if (impl_->getEndEffectorLink().empty() || link_name.empty())
     return false;
@@ -1548,7 +1548,7 @@ bool moveit::planning_interface::MoveGroup::setEndEffectorLink(const std::string
   return true;
 }
 
-bool moveit::planning_interface::MoveGroup::setEndEffector(const std::string &eef_name)
+bool moveit::planning_interface::MoveGroupInterface::setEndEffector(const std::string &eef_name)
 {
   const robot_model::JointModelGroup *jmg = impl_->getRobotModel()->getEndEffector(eef_name);
   if (jmg)
@@ -1556,17 +1556,17 @@ bool moveit::planning_interface::MoveGroup::setEndEffector(const std::string &ee
   return false;
 }
 
-void moveit::planning_interface::MoveGroup::clearPoseTarget(const std::string &end_effector_link)
+void moveit::planning_interface::MoveGroupInterface::clearPoseTarget(const std::string &end_effector_link)
 {
   impl_->clearPoseTarget(end_effector_link);
 }
 
-void moveit::planning_interface::MoveGroup::clearPoseTargets()
+void moveit::planning_interface::MoveGroupInterface::clearPoseTargets()
 {
   impl_->clearPoseTargets();
 }
 
-bool moveit::planning_interface::MoveGroup::setPoseTarget(const Eigen::Affine3d &pose, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setPoseTarget(const Eigen::Affine3d &pose, const std::string &end_effector_link)
 {
   std::vector<geometry_msgs::PoseStamped> pose_msg(1);
   tf::poseEigenToMsg(pose, pose_msg[0].pose);
@@ -1575,7 +1575,7 @@ bool moveit::planning_interface::MoveGroup::setPoseTarget(const Eigen::Affine3d 
   return setPoseTargets(pose_msg, end_effector_link);
 }
 
-bool moveit::planning_interface::MoveGroup::setPoseTarget(const geometry_msgs::Pose &target, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setPoseTarget(const geometry_msgs::Pose &target, const std::string &end_effector_link)
 {
   std::vector<geometry_msgs::PoseStamped> pose_msg(1);
   pose_msg[0].pose = target;
@@ -1584,13 +1584,13 @@ bool moveit::planning_interface::MoveGroup::setPoseTarget(const geometry_msgs::P
   return setPoseTargets(pose_msg, end_effector_link);
 }
 
-bool moveit::planning_interface::MoveGroup::setPoseTarget(const geometry_msgs::PoseStamped &target, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setPoseTarget(const geometry_msgs::PoseStamped &target, const std::string &end_effector_link)
 {
   std::vector<geometry_msgs::PoseStamped> targets(1, target);
   return setPoseTargets(targets, end_effector_link);
 }
 
-bool moveit::planning_interface::MoveGroup::setPoseTargets(const EigenSTL::vector_Affine3d &target, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setPoseTargets(const EigenSTL::vector_Affine3d &target, const std::string &end_effector_link)
 {
   std::vector<geometry_msgs::PoseStamped> pose_out(target.size());
   ros::Time tm = ros::Time::now();
@@ -1604,7 +1604,7 @@ bool moveit::planning_interface::MoveGroup::setPoseTargets(const EigenSTL::vecto
   return setPoseTargets(pose_out, end_effector_link);
 }
 
-bool moveit::planning_interface::MoveGroup::setPoseTargets(const std::vector<geometry_msgs::Pose> &target, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setPoseTargets(const std::vector<geometry_msgs::Pose> &target, const std::string &end_effector_link)
 {
   std::vector<geometry_msgs::PoseStamped> target_stamped(target.size());
   ros::Time tm = ros::Time::now();
@@ -1618,7 +1618,7 @@ bool moveit::planning_interface::MoveGroup::setPoseTargets(const std::vector<geo
   return setPoseTargets(target_stamped, end_effector_link);
 }
 
-bool moveit::planning_interface::MoveGroup::setPoseTargets(const std::vector<geometry_msgs::PoseStamped> &target, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setPoseTargets(const std::vector<geometry_msgs::PoseStamped> &target, const std::string &end_effector_link)
 {
   if (target.empty())
   {
@@ -1632,12 +1632,12 @@ bool moveit::planning_interface::MoveGroup::setPoseTargets(const std::vector<geo
   }
 }
 
-const geometry_msgs::PoseStamped& moveit::planning_interface::MoveGroup::getPoseTarget(const std::string &end_effector_link) const
+const geometry_msgs::PoseStamped& moveit::planning_interface::MoveGroupInterface::getPoseTarget(const std::string &end_effector_link) const
 {
   return impl_->getPoseTarget(end_effector_link);
 }
 
-const std::vector<geometry_msgs::PoseStamped>& moveit::planning_interface::MoveGroup::getPoseTargets(const std::string &end_effector_link) const
+const std::vector<geometry_msgs::PoseStamped>& moveit::planning_interface::MoveGroupInterface::getPoseTargets(const std::string &end_effector_link) const
 {
   return impl_->getPoseTargets(end_effector_link);
 }
@@ -1660,7 +1660,7 @@ inline void transformPose(const tf::Transformer& tf, const std::string &desired_
 }
 }
 
-bool moveit::planning_interface::MoveGroup::setPositionTarget(double x, double y, double z, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setPositionTarget(double x, double y, double z, const std::string &end_effector_link)
 {
   geometry_msgs::PoseStamped target;
   if (impl_->hasPoseTarget(end_effector_link))
@@ -1685,7 +1685,7 @@ bool moveit::planning_interface::MoveGroup::setPositionTarget(double x, double y
   return result;
 }
 
-bool moveit::planning_interface::MoveGroup::setRPYTarget(double r, double p, double y, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setRPYTarget(double r, double p, double y, const std::string &end_effector_link)
 {
   geometry_msgs::PoseStamped target;
   if (impl_->hasPoseTarget(end_effector_link))
@@ -1707,7 +1707,7 @@ bool moveit::planning_interface::MoveGroup::setRPYTarget(double r, double p, dou
   return result;
 }
 
-bool moveit::planning_interface::MoveGroup::setOrientationTarget(double x, double y, double z, double w, const std::string &end_effector_link)
+bool moveit::planning_interface::MoveGroupInterface::setOrientationTarget(double x, double y, double z, double w, const std::string &end_effector_link)
 {
   geometry_msgs::PoseStamped target;
   if (impl_->hasPoseTarget(end_effector_link))
@@ -1732,64 +1732,64 @@ bool moveit::planning_interface::MoveGroup::setOrientationTarget(double x, doubl
   return result;
 }
 
-void moveit::planning_interface::MoveGroup::setPoseReferenceFrame(const std::string &pose_reference_frame)
+void moveit::planning_interface::MoveGroupInterface::setPoseReferenceFrame(const std::string &pose_reference_frame)
 {
   impl_->setPoseReferenceFrame(pose_reference_frame);
 }
 
-const std::string& moveit::planning_interface::MoveGroup::getPoseReferenceFrame() const
+const std::string& moveit::planning_interface::MoveGroupInterface::getPoseReferenceFrame() const
 {
   return impl_->getPoseReferenceFrame();
 }
 
-double moveit::planning_interface::MoveGroup::getGoalJointTolerance() const
+double moveit::planning_interface::MoveGroupInterface::getGoalJointTolerance() const
 {
   return impl_->getGoalJointTolerance();
 }
 
-double moveit::planning_interface::MoveGroup::getGoalPositionTolerance() const
+double moveit::planning_interface::MoveGroupInterface::getGoalPositionTolerance() const
 {
   return impl_->getGoalPositionTolerance();
 }
 
-double moveit::planning_interface::MoveGroup::getGoalOrientationTolerance() const
+double moveit::planning_interface::MoveGroupInterface::getGoalOrientationTolerance() const
 {
   return impl_->getGoalOrientationTolerance();
 }
 
-void moveit::planning_interface::MoveGroup::setGoalTolerance(double tolerance)
+void moveit::planning_interface::MoveGroupInterface::setGoalTolerance(double tolerance)
 {
   setGoalJointTolerance(tolerance);
   setGoalPositionTolerance(tolerance);
   setGoalOrientationTolerance(tolerance);
 }
 
-void moveit::planning_interface::MoveGroup::setGoalJointTolerance(double tolerance)
+void moveit::planning_interface::MoveGroupInterface::setGoalJointTolerance(double tolerance)
 {
   impl_->setGoalJointTolerance(tolerance);
 }
 
-void moveit::planning_interface::MoveGroup::setGoalPositionTolerance(double tolerance)
+void moveit::planning_interface::MoveGroupInterface::setGoalPositionTolerance(double tolerance)
 {
   impl_->setGoalPositionTolerance(tolerance);
 }
 
-void moveit::planning_interface::MoveGroup::setGoalOrientationTolerance(double tolerance)
+void moveit::planning_interface::MoveGroupInterface::setGoalOrientationTolerance(double tolerance)
 {
   impl_->setGoalOrientationTolerance(tolerance);
 }
 
-void moveit::planning_interface::MoveGroup::rememberJointValues(const std::string &name)
+void moveit::planning_interface::MoveGroupInterface::rememberJointValues(const std::string &name)
 {
   rememberJointValues(name, getCurrentJointValues());
 }
 
-bool moveit::planning_interface::MoveGroup::startStateMonitor(double wait)
+bool moveit::planning_interface::MoveGroupInterface::startStateMonitor(double wait)
 {
   return impl_->startStateMonitor(wait);
 }
 
-std::vector<double> moveit::planning_interface::MoveGroup::getCurrentJointValues()
+std::vector<double> moveit::planning_interface::MoveGroupInterface::getCurrentJointValues()
 {
   robot_state::RobotStatePtr current_state;
   std::vector<double> values;
@@ -1798,14 +1798,14 @@ std::vector<double> moveit::planning_interface::MoveGroup::getCurrentJointValues
   return values;
 }
 
-std::vector<double> moveit::planning_interface::MoveGroup::getRandomJointValues()
+std::vector<double> moveit::planning_interface::MoveGroupInterface::getRandomJointValues()
 {
   std::vector<double> r;
   impl_->getJointModelGroup()->getVariableRandomPositions(impl_->getJointStateTarget().getRandomNumberGenerator(), r);
   return r;
 }
 
-geometry_msgs::PoseStamped moveit::planning_interface::MoveGroup::getRandomPose(const std::string &end_effector_link)
+geometry_msgs::PoseStamped moveit::planning_interface::MoveGroupInterface::getRandomPose(const std::string &end_effector_link)
 {
   const std::string &eef = end_effector_link.empty() ? getEndEffectorLink() : end_effector_link;
   Eigen::Affine3d pose;
@@ -1830,7 +1830,7 @@ geometry_msgs::PoseStamped moveit::planning_interface::MoveGroup::getRandomPose(
   return pose_msg;
 }
 
-geometry_msgs::PoseStamped moveit::planning_interface::MoveGroup::getCurrentPose(const std::string &end_effector_link)
+geometry_msgs::PoseStamped moveit::planning_interface::MoveGroupInterface::getCurrentPose(const std::string &end_effector_link)
 {
   const std::string &eef = end_effector_link.empty() ? getEndEffectorLink() : end_effector_link;
   Eigen::Affine3d pose;
@@ -1854,7 +1854,7 @@ geometry_msgs::PoseStamped moveit::planning_interface::MoveGroup::getCurrentPose
   return pose_msg;
 }
 
-std::vector<double> moveit::planning_interface::MoveGroup::getCurrentRPY(const std::string &end_effector_link)
+std::vector<double> moveit::planning_interface::MoveGroupInterface::getCurrentRPY(const std::string &end_effector_link)
 {
   std::vector<double> result;
   const std::string &eef = end_effector_link.empty() ? getEndEffectorLink() : end_effector_link;
@@ -1882,114 +1882,114 @@ std::vector<double> moveit::planning_interface::MoveGroup::getCurrentRPY(const s
   return result;
 }
 
-const std::vector<std::string>& moveit::planning_interface::MoveGroup::getActiveJoints() const
+const std::vector<std::string>& moveit::planning_interface::MoveGroupInterface::getActiveJoints() const
 {
   return impl_->getJointModelGroup()->getActiveJointModelNames();
 }
 
-const std::vector<std::string>& moveit::planning_interface::MoveGroup::getJoints() const
+const std::vector<std::string>& moveit::planning_interface::MoveGroupInterface::getJoints() const
 {
   return impl_->getJointModelGroup()->getJointModelNames();
 }
 
-unsigned int moveit::planning_interface::MoveGroup::getVariableCount() const
+unsigned int moveit::planning_interface::MoveGroupInterface::getVariableCount() const
 {
   return impl_->getJointModelGroup()->getVariableCount();
 }
 
-robot_state::RobotStatePtr moveit::planning_interface::MoveGroup::getCurrentState()
+robot_state::RobotStatePtr moveit::planning_interface::MoveGroupInterface::getCurrentState()
 {
   robot_state::RobotStatePtr current_state;
   impl_->getCurrentState(current_state);
   return current_state;
 }
 
-void moveit::planning_interface::MoveGroup::rememberJointValues(const std::string &name, const std::vector<double> &values)
+void moveit::planning_interface::MoveGroupInterface::rememberJointValues(const std::string &name, const std::vector<double> &values)
 {
   remembered_joint_values_[name] = values;
 }
 
-void moveit::planning_interface::MoveGroup::forgetJointValues(const std::string &name)
+void moveit::planning_interface::MoveGroupInterface::forgetJointValues(const std::string &name)
 {
   remembered_joint_values_.erase(name);
 }
 
-void moveit::planning_interface::MoveGroup::allowLooking(bool flag)
+void moveit::planning_interface::MoveGroupInterface::allowLooking(bool flag)
 {
   impl_->allowLooking(flag);
 }
 
-void moveit::planning_interface::MoveGroup::allowReplanning(bool flag)
+void moveit::planning_interface::MoveGroupInterface::allowReplanning(bool flag)
 {
   impl_->allowReplanning(flag);
 }
 
-std::vector<std::string> moveit::planning_interface::MoveGroup::getKnownConstraints() const
+std::vector<std::string> moveit::planning_interface::MoveGroupInterface::getKnownConstraints() const
 {
   return impl_->getKnownConstraints();
 }
 
-moveit_msgs::Constraints moveit::planning_interface::MoveGroup::getPathConstraints() const
+moveit_msgs::Constraints moveit::planning_interface::MoveGroupInterface::getPathConstraints() const
 {
    return impl_->getPathConstraints();
 }
 
-bool moveit::planning_interface::MoveGroup::setPathConstraints(const std::string &constraint)
+bool moveit::planning_interface::MoveGroupInterface::setPathConstraints(const std::string &constraint)
 {
   return impl_->setPathConstraints(constraint);
 }
 
-void moveit::planning_interface::MoveGroup::setPathConstraints(const moveit_msgs::Constraints &constraint)
+void moveit::planning_interface::MoveGroupInterface::setPathConstraints(const moveit_msgs::Constraints &constraint)
 {
   impl_->setPathConstraints(constraint);
 }
 
-void moveit::planning_interface::MoveGroup::clearPathConstraints()
+void moveit::planning_interface::MoveGroupInterface::clearPathConstraints()
 {
   impl_->clearPathConstraints();
 }
 
-void moveit::planning_interface::MoveGroup::setConstraintsDatabase(const std::string &host, unsigned int port)
+void moveit::planning_interface::MoveGroupInterface::setConstraintsDatabase(const std::string &host, unsigned int port)
 {
   impl_->initializeConstraintsStorage(host, port);
 }
 
-void moveit::planning_interface::MoveGroup::setWorkspace(double minx, double miny, double minz, double maxx, double maxy, double maxz)
+void moveit::planning_interface::MoveGroupInterface::setWorkspace(double minx, double miny, double minz, double maxx, double maxy, double maxz)
 {
   impl_->setWorkspace(minx, miny, minz, maxx, maxy, maxz);
 }
 
-void moveit::planning_interface::MoveGroup::setPlanningTime(double seconds)
+void moveit::planning_interface::MoveGroupInterface::setPlanningTime(double seconds)
 {
   impl_->setPlanningTime(seconds);
 }
 
-double moveit::planning_interface::MoveGroup::getPlanningTime() const
+double moveit::planning_interface::MoveGroupInterface::getPlanningTime() const
 {
   return impl_->getPlanningTime();
 }
 
-void moveit::planning_interface::MoveGroup::setSupportSurfaceName(const std::string &name)
+void moveit::planning_interface::MoveGroupInterface::setSupportSurfaceName(const std::string &name)
 {
   impl_->setSupportSurfaceName(name);
 }
 
-const std::string& moveit::planning_interface::MoveGroup::getPlanningFrame() const
+const std::string& moveit::planning_interface::MoveGroupInterface::getPlanningFrame() const
 {
   return impl_->getRobotModel()->getModelFrame();
 }
 
-bool moveit::planning_interface::MoveGroup::attachObject(const std::string &object, const std::string &link)
+bool moveit::planning_interface::MoveGroupInterface::attachObject(const std::string &object, const std::string &link)
 {
   return attachObject(object, link, std::vector<std::string>());
 }
 
-bool moveit::planning_interface::MoveGroup::attachObject(const std::string &object, const std::string &link, const std::vector<std::string> &touch_links)
+bool moveit::planning_interface::MoveGroupInterface::attachObject(const std::string &object, const std::string &link, const std::vector<std::string> &touch_links)
 {
   return impl_->attachObject(object, link, touch_links);
 }
 
-bool moveit::planning_interface::MoveGroup::detachObject(const std::string &name)
+bool moveit::planning_interface::MoveGroupInterface::detachObject(const std::string &name)
 {
   return impl_->detachObject(name);
 }
